+++ example/index.js
@@ -1,4 +1,0 @@
-
-const seeFetch = require('../dist/see-fetch');
-
-require('../exec')(seeFetch);