+++ src/data.js
@@ -1,9 +1,0 @@
-
-'use strict';
-
-module.exports = {
-    // application options
-    options: {},
-    // environment, default is 0
-    env: 0
-};