+++ src/get_env.js
@@ -1,13 +1,6 @@
-
-'use strict';
-
-var data = require('./data');
+import share from './share';
 
 /**
- * get current environment
- *
- * @returns {Object.<string, *>|null|*|number}
+ * Get current environment.
  */
-module.exports = () => {
-    return data.env;
-};
+export default () => share.env;