+++ src/post_fetch_handle.js
@@ -1,15 +1,9 @@
+import postHandle from './post_handle';
+import setting from './setting';
 
-const postHandle = require('./post_handle');
-const setting = require('./setting');
+export default (name, params) => res => {
+  // Has error.
+  if (res[setting.errorField]) return res;
 
-module.exports = (name, ultimateReqData) => {
-    return res => {
-        // has error
-        if (res[setting.errorField]) return res;
-
-        // post handle
-        postHandle(res, ultimateReqData, name);
-
-        return res;
-    };
+  return postHandle(res, params, name);
 };