+++ src/send.js
@@ -1,102 +1,130 @@
-
-'use strict';
-
-const setting = require('./setting');
-const data = require('./data');
-const logger = require('./util/logger');
-const makeSearch = require('./util/make_search');
-const makeUrlSearch = require('./util/make_url_search');
-const fetchHandle = require('./fetch_handle');
-const postFetchHandle = require('./post_fetch_handle');
-const postHandle = require('./post_handle');
+import setting from './setting';
+import share from './share';
+import { info, error } from './util/logger';
+import makeSearch from './util/make_search';
+import makeUrlSearchParams from './util/make_url_search_params';
+import fetchHandle from './fetch_handle';
+import postFetchHandle from './post_fetch_handle';
+import postHandle from './post_handle';
 
 /**
- * send a request
+ * Make a request.
  *
- * @param name Defined request name
- * @param reqData Request data
+ * @param name Defined request name.
+ * @param params Request params.
  */
-module.exports = (name, reqData) => {
-    if (!name) {
-        logger.throwError(`name '${name}' is not defined.`);
-        return;
+export default function(name, params) {
+  if (!name) return;
+
+  // Current options.
+  const options = share.optionsCollection[name];
+  // Common options.
+  const commonOptions = share.optionsCollection.common || {};
+
+  if (!options) {
+    error(`name '${name}' is not configured.`);
+    return;
+  }
+
+  // Index to select item.
+  const index = share.env;
+
+  // Http method, default is `GET`.
+  const method = (options.method && options.method[index]) || 'get';
+
+  // Stringify request params.
+  const stringify = (options.stringify && options.stringify[index]) || !1;
+
+  // Fetch options.
+  const settings = (options.settings && options.settings[index]) || {};
+
+  // url
+  const url = (options.url && options.url[index]) || '';
+
+  // Request keys.
+  const requestKeys = (options.requestKeys && options.requestKeys[index]) || {};
+
+  // Pre handle.
+  const preHandle = options.preHandle && options.preHandle[index];
+
+  const commonPreHandle = commonOptions.preHandle && commonOptions.preHandle[index];
+
+  // implement
+  const implement = options.implement && options.implement[index];
+
+  // Real request params.
+  let realParams = Object.assign({}, params || {});
+
+  // Request keys mapping handling.
+  Object.keys(realParams).forEach(key => {
+    const newKey = requestKeys[key];
+    if (newKey && typeof newKey === 'string') {
+      // Make a new key.
+      realParams[newKey] = realParams[key];
+      // Delete old key.
+      delete realParams[key];
     }
+  });
 
-    // current option
-    var option = data.options[name];
-    // common option
-    var commonOption = data.options['common'] || {};
+  // Pre handling.
+  if (commonPreHandle) {
+    const result = commonPreHandle(realParams);
 
-    if (!option) {
-        logger.throwError(`name '${name}' is not configured.`);
-        return;
-    }
+    // If return a new object, use it.
+    if (result) realParams = result;
+  }
+  if (preHandle) {
+    const result = preHandle(realParams);
 
-    // index to select item
-    var index = data.env;
+    // If return a new object, use it.
+    if (result) realParams = result;
+  }
 
-    // http method, default is GET
-    var method = option.method && option.method[index] || 'get';
-    // stringify request data
-    var stringify = option.stringify && option.stringify[index] || !1;
-    // fetch options
-    var settings = option.settings && option.settings[index] || {};
-    // url
-    var url = option.url && option.url[index] || '';
-    // request keys
-    var requestKeys = option.requestKeys && option.requestKeys[index] || {};
-    // pre handle
-    var preHandle = option.preHandle && option.preHandle[index];
-    var commonPreHandle = commonOption.preHandle && commonOption.preHandle[index];
-    // implement
-    var implement = option.implement && option.implement[index];
-
-    // ultimate request data after requestKeys mapping
-    var ultimateReqData = Object.assign({}, reqData || {});
-    for (var ultimateReqDataAttr in ultimateReqData) {
-        if (ultimateReqData.hasOwnProperty(ultimateReqDataAttr) && requestKeys[ultimateReqDataAttr]) {
-            // make a new key
-            ultimateReqData[requestKeys[ultimateReqDataAttr]] = ultimateReqData[ultimateReqDataAttr];
-            // delete old key
-            delete ultimateReqData[ultimateReqDataAttr];
+  // Custom implement.
+  if (implement) {
+    return new Promise(resolve => {
+      const callback = result => {
+        if (setting.debug) {
+          info(`custom fetch implement for '${name}', and request params is:`, realParams);
+          info(`result for '${name}' is:`, result);
         }
-    }
 
-    // pre handle
-    commonPreHandle && commonPreHandle(ultimateReqData);
-    preHandle && preHandle(ultimateReqData);
+        resolve(postHandle(result, realParams, name));
+      };
 
-    // custom implement
-    if (implement) {
-        return new Promise((resolve) => {
-            implement(result => {
-                if (setting.debug) {
-                    logger.info(`Custom implement fetch for "${name}", and request data is:`);
-                    console.log(ultimateReqData);
-                    logger.info(`result for "${name}" is:`);
-                    console.log(result);
-                }
+      // Use callback
+      const promise = implement(result => {
+        callback(result);
+      }, !stringify ? realParams : JSON.stringify(realParams));
 
-                // post handle
-                postHandle(result, ultimateReqData, name);
-
-                resolve(result);
-            }, !stringify ? ultimateReqData : JSON.stringify(ultimateReqData));
+      // Return a Promise
+      if (promise && promise instanceof Promise)
+        promise.then(result => {
+          callback(result);
         });
+    });
+  } else {
+    settings.method = method;
+    if (method === 'get' || method === 'GET') {
+      const newUrl = url + (url.indexOf('?') < 0 ? '?' : '&') + makeSearch(realParams);
+      return fetch(newUrl, settings)
+        .then(fetchHandle)
+        .then(postFetchHandle(name, realParams));
     }
-    else {
-        settings.method = method;
-        if (method === 'get' || method === 'GET') {
-            let newUrl = url + (url.indexOf('?') < 0 ? '?' : '&') + makeSearch(ultimateReqData);
-            return fetch(newUrl, settings).then(fetchHandle).then(postFetchHandle(name, ultimateReqData));
-        }
 
-        if (stringify) {
-            settings.body = JSON.stringify(ultimateReqData);
-            return fetch(url, settings).then(fetchHandle).then(postFetchHandle(name, ultimateReqData));
-        }
+    settings.body = stringify ? JSON.stringify(realParams) : makeUrlSearchParams(realParams);
 
-        settings.body = makeUrlSearch(ultimateReqData);
-        return fetch(url, settings).then(fetchHandle).then(postFetchHandle(name, ultimateReqData));
+    if (method !== 'get' && method !== 'GET' && method !== 'head' && method !== 'HEAD') {
+      if (!settings.headers) settings.headers = {};
+
+      if (!settings.headers['Content-Type'])
+        settings.headers['Content-Type'] = stringify
+          ? 'application/json'
+          : 'application/x-www-form-urlencoded;charset=UTF-8';
     }
-};
+
+    return fetch(url, settings)
+      .then(fetchHandle)
+      .then(postFetchHandle(name, realParams));
+  }
+}