+++ src/setting.js
@@ -1,7 +1,6 @@
-
-module.exports = {
-    // error field when response statusCode
-    errorField: 'error',
-    // whether current mode is debug
-    debug: !0
+export default {
+  // Error field when response status is `3XX, 4XX, 5XX`.
+  errorField: 'error',
+  // Whether current mode is debug.
+  debug: !0,
 };