+++ src/util/logger.js
@@ -1,22 +1,21 @@
+const prefix = 'see-fetch: ';
 
-'use strict';
+export const log = (str, ...rest) => {
+  console.log(prefix + str, ...rest);
+};
 
-var prefix = 'see-fetch: ';
+export const info = (str, ...rest) => {
+  console.info(prefix + str, ...rest);
+};
 
-module.exports = {
-    log: (str) => {
-        console.log(prefix + str);
-    },
-    info: (str) => {
-        console.info(prefix + str);
-    },
-    warn: (str) => {
-        console.warn(prefix + str);
-    },
-    error: (str) => {
-        console.error(prefix + str);
-    },
-    throwError: (str) => {
-        throw new Error(prefix + str);
-    }
+export const warn = (str, ...rest) => {
+  console.warn(prefix + str, ...rest);
+};
+
+export const error = (str, ...rest) => {
+  console.error(prefix + str, ...rest);
+};
+
+export const throwError = str => {
+  throw new Error(prefix + str);
 };