+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.0 [2018-09-06]
+
+- Use `microbundle` to build project.
+
+## 0.1.1 [2018-08-27]
+
+- `implement` can return `Promise`.
+
 ## 0.1.0 [2018-07-24]
 
 - Remove `implementDelay` option.